requirements.txt
cinemagoer==2023.5.1
greenlet==3.1.1
IMDbPY==2022.7.9
lxml==5.3.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
